requirements.txt
flask
gunicorn
sentence-transformers
faiss-cpu
groq
pytesseract
Pillow
easyocr
torch
torchvision
numpy